test: isolate hosted message gateway routing coverage
