+++ src/agents/tools/gateway.test.ts
@@ -1,11 +1,16 @@
-// Register Gateway mocks before importing runtime dependencies.
-import "./gateway.test-helpers.js";
 import { expectDefined } from "@openclaw/normalization-core";
-import { describe, expect, it, vi } from "vitest";
+import { afterAll, beforeEach, describe, expect, it, vi } from "vitest";
 import { verifyAgentRuntimeIdentityToken } from "../../gateway/agent-runtime-identity-token.js";
 import type { CallGatewayOptions } from "../../gateway/call.js";
+import type { GatewayRequestContext } from "../../gateway/server-methods/types.js";
+import {
+  claimAgentRunDelegatedAuthority,
+  releaseAgentRunDelegatedAuthority,
+  type AgentRunDelegatedAuthority,
+} from "../../infra/agent-run-registry.js";
 import { createEmptyPluginRegistry } from "../../plugins/registry-empty.js";
 import { setActivePluginRegistry } from "../../plugins/runtime.js";
+import { createOperationalRunInstanceRef } from "../admitted-run-context.js";
 import { withGatewayToolCallerIdentity } from "./gateway-caller-context.js";
 import {
   callGatewayTool,
@@ -13,15 +18,109 @@ import {
   resolveGatewayOptions,
   resolveMessageActionAgentRuntimeIdentityToken,
 } from "./gateway.js";
-import {
-  capturedGatewayCall,
-  installGatewayToolTestHooks,
-  mocks,
-  testGatewayCaller,
-} from "./gateway.test-helpers.js";
+
+const mocks = vi.hoisted(() => ({
+  callGateway: vi.fn(),
+  configState: {
+    value: {} as Record<string, unknown>,
+  },
+  deviceIdentity: {
+    deviceId: "agent-tool-device",
+    publicKeyPem: "public-key",
+    privateKeyPem: "private-key",
+  },
+  persistedDeviceIdentity: undefined as
+    | {
+        deviceId: string;
+        publicKeyPem: string;
+        privateKeyPem: string;
+      }
+    | null
+    | undefined,
+  deviceIdentityError: undefined as Error | undefined,
+}));
+const testDelegatedAuthorities: AgentRunDelegatedAuthority[] = [];
+
+function releaseTestDelegatedAuthorities(): void {
+  for (const authority of testDelegatedAuthorities.splice(0)) {
+    releaseAgentRunDelegatedAuthority(authority);
+  }
+}
+vi.mock("../../config/config.js", () => ({
+  getRuntimeConfig: () => mocks.configState.value,
+  resolveGatewayPort: () => 18789,
+}));
+vi.mock("../../gateway/call.js", () => ({
+  callGateway: (...args: unknown[]) => mocks.callGateway(...args),
+}));
+vi.mock("../../infra/device-identity.js", () => ({
+  loadDeviceIdentityIfPresent: () =>
+    mocks.persistedDeviceIdentity === undefined
+      ? mocks.deviceIdentity
+      : mocks.persistedDeviceIdentity,
+  loadOrCreateDeviceIdentity: () => {
+    if (mocks.deviceIdentityError) {
+      throw mocks.deviceIdentityError;
+    }
+    return mocks.deviceIdentity;
+  },
+}));
+
+function capturedGatewayCall(): CallGatewayOptions {
+  expect(mocks.callGateway).toHaveBeenCalledTimes(1);
+  const call = mocks.callGateway.mock.calls[0];
+  if (!call) {
+    throw new Error("expected callGateway to be called");
+  }
+  return call[0] as CallGatewayOptions;
+}
+
+function testGatewayCaller(
+  identity: Omit<
+    NonNullable<Parameters<typeof withGatewayToolCallerIdentity>[0]>,
+    "operationalRunInstance"
+  >,
+): NonNullable<Parameters<typeof withGatewayToolCallerIdentity>[0]> {
+  const operationalRunInstance = createOperationalRunInstanceRef("run-gateway-tool-test");
+  testDelegatedAuthorities.push(claimAgentRunDelegatedAuthority(operationalRunInstance));
+  const context = { getRuntimeConfig: () => mocks.configState.value } as GatewayRequestContext;
+  return {
+    gatewayContextResolver: () => context,
+    ...identity,
+    operationalRunInstance,
+  };
+}
 
 describe("gateway tool defaults", () => {
-  installGatewayToolTestHooks();
+  const envSnapshot = {
+    openclaw: process.env.OPENCLAW_GATEWAY_TOKEN,
+    gatewayUrl: process.env.OPENCLAW_GATEWAY_URL,
+  };
+
+  beforeEach(() => {
+    releaseTestDelegatedAuthorities();
+    mocks.callGateway.mockReset();
+    mocks.deviceIdentityError = undefined;
+    mocks.persistedDeviceIdentity = undefined;
+    mocks.configState.value = {};
+    setActivePluginRegistry(createEmptyPluginRegistry());
+    delete process.env.OPENCLAW_GATEWAY_TOKEN;
+    delete process.env.OPENCLAW_GATEWAY_URL;
+  });
+
+  afterAll(() => {
+    releaseTestDelegatedAuthorities();
+    if (envSnapshot.openclaw === undefined) {
+      delete process.env.OPENCLAW_GATEWAY_TOKEN;
+    } else {
+      process.env.OPENCLAW_GATEWAY_TOKEN = envSnapshot.openclaw;
+    }
+    if (envSnapshot.gatewayUrl === undefined) {
+      delete process.env.OPENCLAW_GATEWAY_URL;
+    } else {
+      process.env.OPENCLAW_GATEWAY_URL = envSnapshot.gatewayUrl;
+    }
+  });
 
   it("leaves url undefined so callGateway can use config", () => {
     const opts = resolveGatewayOptions();
